requirements.txt
flask
requests
waitress
pyjwt
paho-mqtt